Varios roles unido

+++ app/Http/Controllers/AulaController.php
@@ -10,6 +10,9 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 use SebastianBergmann\Environment\Console;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Requests\AulaCreateRequest;
+use App\Http\Requests\AulaEditRequest;
 use Alert;
 
 class AulaController extends Controller
@@ -38,7 +41,7 @@ class AulaController extends Controller
             return view('admin.aulas.index', compact('aulas'))->with('tipo', "reservadas");
 
         }elseif($request->tipo === "admin"){
-            
+            abort_if(Gate::denies('aulaR_index'), 403);
             $aulas = DB::table('solicitudes')
             ->join('docmaterias', 'solicitudes.docmateria_id', '=', 'docmaterias.id')
             ->join('materias', 'docmaterias.materia', '=', 'materias.id')
@@ -50,7 +53,8 @@ class AulaController extends Controller
             return view('admin.aulasR.index', compact('aulas'))->with('tipo', "admin");
 
         }
- 
+        
+        abort_if(Gate::denies('aula_index'), 403);
         $aulas =  DB::table('aulas')
         ->join('sectors', 'aulas.sector', '=', 'sectors.id')
         ->select('aulas.*','sectors.nombre')
@@ -68,7 +72,9 @@ class AulaController extends Controller
      */
     public function create()
     {
-        //
+         abort_if(Gate::denies('aula_create'), 403);
+         $sectors = Sector::all()->pluck('nombre', 'id');
+        return view('admin.aulas.create');
     }
 
     /**
@@ -77,9 +83,11 @@ class AulaController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
+    
     public function store(Request $request)
     {      
-        $newAula= new Aula();
+        abort_if(Gate::denies('aula_create'), 403);
+            $newAula= new Aula();
    
             $newAula->codigo = $request->codigo;
             $newAula->num_aula = $request->num_aula;
@@ -94,7 +102,7 @@ class AulaController extends Controller
                 return redirect()->back();
             }else{ 
                 
-                return back()->withErrors([
+                return back()->withInput()->withErrors([
                     'message' => 'Error, el codigo o numero de aula ingresado ya existe'
                 ]);
             }
@@ -106,6 +114,7 @@ class AulaController extends Controller
 
     public function delete(Request $request, $aulaId)
     {
+        abort_if(Gate::denies('aula_destroy'), 403);
         function debug_to_console($data) {
             $output = $data;
             if (is_array($output))
@@ -147,7 +156,7 @@ class AulaController extends Controller
     }
     public function deleteReservadas(Request $request, $reservaId)
     {
-        
+        abort_if(Gate::denies('aula_destroy'), 403);
         $reserva = Solicitud::find($reservaId);
         $reserva->delete();
         return redirect()->back();
@@ -184,12 +193,25 @@ class AulaController extends Controller
      */
     public function update(Request $request, $aulaId)
     {
+        abort_if(Gate::denies('aula_edit'), 403);
         $aula = Aula::find($aulaId);
         $aula->num_aula = $request->num_aula;
         $aula->capacidad = $request->capacidad;
         $aula->sector = $request->sector;
         $aula->estado = $request->estado;
-        $aula->save();
+       
+
+        $aula2 = Aula::where('num_aula', $request->num_aula)->first();
+            if(empty($aula2)){    
+                 $aula->save();
+                return redirect()->back();
+            }else{ 
+                
+                return back()->withInput()->withErrors([
+                    'message' => 'Error, El numero de aula ingresado ya existe'
+                ]);
+            }
+           
 
        return redirect()->back();
     }
